PewPew improvements
